refactor(nGyu) : Utils 분리
