+++ racingcar-java/src/test/java/step3/RacingTest.java
@@ -2,13 +2,11 @@ package step3;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import step3.domain.Car;
 import step3.domain.Cars;
-import step3.view.Utils;
+import step3.view.OutputView;
 
 import java.io.ByteArrayInputStream;
 import java.io.InputStream;
-import java.util.Arrays;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
@@ -16,7 +14,7 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 @DisplayName("레이싱 테스트")
 class RacingTest {
     Racing racing = new Racing();
-    Utils.OutputView outputView = new Utils.OutputView();
+    OutputView outputView = new OutputView();
 
     @Test
     @DisplayName("입력값 넣기 테스트")
